Restore original layout for core utilities
